[Breaking] Move generateCode from org.jetbrains.kotlinx.dataframe.codeGen to org.jetbrains.kotlinx.dataframe.api
Having to import this function in notebooks is bothersome. Api package is imported by default. Adding codeGen to default imports would add too much things.
